Extract Http mocking setup to extension method
